requirements/local.txt
-r ./base.txt

# General
# ------------------------------------------------------------------------------
Werkzeug==0.16.0 # https://github.com/pallets/werkzeug
ipdb==0.12.3  # https://github.com/gotcha/ipdb
Sphinx==2.3.0  # https://github.com/sphinx-doc/sphinx
psycopg2==2.8.3 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.9  # https://github.com/PyCQA/flake8
coverage==5.0  # https://github.com/nedbat/coveragepy
black==19.10b0  # https://github.com/ambv/black
pylint-django==2.0.13  # https://github.com/PyCQA/pylint-django
pre-commit==1.20.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==2.12.0
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
django-coverage-plugin==1.6.0
pytest-django==3.7.0